feat: add MagnifyImage component
